requirements.txt
django>=3.2.12,<3.3.0
djangorestframework>=3.13.1,<3.14.0
flake8>=3.6.0,<=3.7.0
python-decouple>=3.6
psycopg2>=2.9.3,<2.10.0
Pillow>=9.1.0,<9.2.0